Support7Demos: Prevent crashes when connecting to sample routes
Since the app lacked a runtime overlay permission, it could not make
a screen which is drawn on top of other apps. Therefore in API >= 23,
whenever a user tried to connect to any of the sample routes, the app
crashed with a permission error. This CL makes the app request the
permission at runtime.

Bug: 63313783
Test: In API 23-26 devices, the app did not crash.
Change-Id: Idb85e3685f4fefa657f50c5b1cd2daa536de535f