MediaRouter: Fix possible IllegalArgumentException
In GlobalMediaRouter.updateSelectedRouteIfNeeded(), when the original
selected route is unselectable, we will set mSelectedRoute to null, and
then set it to a fallback route. However, this would create a short window
that mSelectedRoute is null, which is conflicting with our assumption
that mSelectedRoute should not be null, as long as the global router is
created.

Directly sets the selected route to a fallback route to prevent this
situation to happen.

Test: ./gradlew support-mediarouter-v7:connectedCheck
Test: ./gradlew :support-v7-demos:installDebug
Bug: 63630303
Change-Id: I0498e750f9389ba9bcbacf2c6d27d57fa38e76af