Fix InvalidationTracker for custom database
InvalidationTracker queries the database when something changes, but
this can cause exception when the database is closed during the query is
being run. We prevent this by catching the exception, but that's not
enough for custom database implementation as it might throw some types
of exception we don't know about.

This CL introduces a close lock between RoomDatabase and
InvalidationTracker. As far as the database is closed with
RoomDatabase.close(), InvalidationTracker can tell that it is being
closed and stop querying it.

We still need to catch IllegalStateException and SQLiteException because
some tests have to close the underlying SupportSQLiteDatabase bypassing
RoomDatabase and currently there's no way InvalidationTracker can hook
into it.

Bug: 63162311
Test: CustomDatbaseTest
Change-Id: Iece7f95eac886287ea09d69ed44cba786340872b