MediaRouter: Fix race condition between selecting routes
The bluetooth route updating in the framework MediaRouter is notified
from AIDL, which would happen in a random thread. And in support
library, the synchronizing of selected route is done with runnable
posting. Therefore, there's a possibility that the bluetooth route is
added in-between sync runnable is posted but not yet executed, which
might cause incorrect synchronization of the selected route that
overrides the newly selected bluetooth route.

Removes sync messages for route selected when a system route is selected
to fix the issue.

Test: ./gradlew support-mediarouter-v7:connectedCheck
Test: ./gradlew :support-v7-demos:installDebug
Bug: 63907134
Change-Id: I733b00c05722b2a29cb72a2eceb8dcf0db4c8e4f