customtabs: Add the validateRelationship() service call.
Websites can declare relationships with native apps, through Digital
Asset Links. This adds an API call to validate that such a relation has
been declared.
This is used to grant special rights to the calling application. For
instance, this is currently used for the postMessage() implementation.

Validation is asynchronous, per-session, and cached by the browser for
the session lifetime.

Test: In Chrome, and unit tests.
Change-Id: If78872a5bf67ccf0a9375703e5a17c3c24516e62