Fix previously uncallable methods in AccessibilityEventCompat
AccessibilityEventCompat is final and it has a private constructor.
We had getAction, setAction, getMovementGranularity, and setMovementGranularity
methods that were non static, thus it was impossible for anyone to call.

This CL adds 26.0.0.ignore file that suppresses the change from non-static
methods to static since in this instance it is an ok change.

Bug: 64378281
Test: ./gradlew assembleDebug checkApi
Change-Id: I2a524ab3ccccc9e0dbcc4bfb96e033a2dbfbb1dc