leanback: disable outline clipping on low power chipset
View outline based rounded corner is the bottleneck on low power
chipset. Adding OEM customization to disable outline clipping. In
the future we will try to implement rounded corner using Path and
Shape in ImageCardView and teach developer how to write a custom
card view doing so.

Bug: 64910200
Test: Manually run sample on low power device with customization apk
and on nexus player. Run ListRowPresenterTest on both devices.

Change-Id: I63bbf4e8764b7614334c2bd0964c148d793eb35f