Catch Resource.NotFoundExceptions for errors
When printing error messages in Navigation, we
use getResourceName() to get a more recognizable
value. However, this fails to account for the possibility
that developers dynamically build NavDestinations that
may not have predefined resource ids. By catching the
Resource.NotFoundException, we ensure that developers get
useful error messages in all cases.

Test: new tests pass
BUG: 65122519
Change-Id: Ieaf70900172715bc48bbcdb8a83468f560d2a8e7