Move createDeepLink methods to NavDeepLinkBuilder
Instead of requiring a full NavController instance
to construct deep links, move the functionality into
a new NavDeepLinkBuilder so that it is useable
in Services/BroadcastReceivers where developers might
want to construct PendingIntents for notifications,
for example.

Also includes a default implementation of the
NavigatorProvider interface in SimpleNavigatorProvider,
removing code from NavController and making it possible
to reuse it to implement the new functionality.

Test: tests pass
BUG: 36817937
Change-Id: I0c19069b7755da8783073486f3b6ae75d4407145