Limit characters in table/column names
SQL is very free on column/table names as long as you put them
inside quotes. This is weird, hardly ever useful and makes code-gen harder.

Instead, we will just limit the set.

This CL also fixes a bug in invalidation tracker where we would
not properly observe tables if it has a space in the name.

Bug: 64749111
Test: SqliteParserTest, EntityProcessorTest, FunnyNamedDaoTest
Change-Id: I6322ee6b3b2c7c74dadc2731032969e52cfbbd98