Handle re-entry in invalidation tracker
When an instant task executor is used, InvalidationTracker would
skip refresh if it is inside transaction but leave the pending
refresh flag on, which would mean the next endTransaction would
not do anything.

This CL fixes that issue by not triggering the invalidation
tracker if current thread is still in a transaction.

It also changes InvalidationTracker to handle re-entry properly.

Bug: 65471397
Test: RxJava2WithInstantTaskExecutorTest
Change-Id: I97d1d5282080593afa97477a1e9b91f91535c04e