Validate indices in migrations
This CL improves the database migration to include index validation.

Even though we can extract more information about the index like the
order, since Room APIs do not allow it, we do not verify them.

The pragma was added in 2015 so if we cannot read index value,
we skip index validation.

Bug: 63132683
Test: MigrationTest, TableInfoTest
Change-Id: I01f35be08232f6a6bb22766553e48f7b177a21e4