leanback: fix the requestAudioFocus nightmare
In b/17851959, Search fragment was made to requestAudioFocus
to duck background music. Then later further change is made to
stop listening when lose focus. This causes recognition aborted when
GSA for TV 3.0 starts to duck the music too.

The original bug really should be fixed in GSA since we should only
have one entity to requestAudioFocus. This CL will cause
a regression of audio not ducking on devices not yet updated
to GSA 3.0 but it should be resolved soon and app will take time
to recompile with not yet released support lib.

Test: manually tested sample app
Bug: 17851959
Bug: 67119344

Change-Id: I86a087e5a99b141fb14c1626c822fdefd667899a