Support @Transaction in @Query methods.
This CL adds support for using @Transaction in @Query methods.

There are 2 major use cases:
a) query result is big so may not fit in 1 cursor window.
b) query has @Relation fields which are queried separately so
may become inconsistent. For this case, we are printing a warning
to educate the developer and letting them choose if they want to
do it.

Bug: 66011356
Bug: 65112315
Test: QueryTransactionTest, DaoProcessorTest, QueryMethodProcessorTest

Change-Id: I128ce7504f8d042bf427d8eb3257654ce4dda1ff