Synchronous testing for ForegroundProcessor.
There are two main changes in this CL:

1. The use of a core-testing's CountingTaskExecutorRule.  LiveData runs
on its own Executor; this adds a delegate to that Executor to allow us to
observe and drain tasks to completion.  This allows synchronous LiveData
behavior (used when the dependent task is observed to be unblocked).

2. A new class, SynchronousExecutorService, which does not spawn any new
threads.  Instead, when a job is submitted to it, it runs it
synchronously on the same thread.  Note that many operations are not
(yet?) implemented in this class and will simply throw an Exception.

Test: Updated tests.
Change-Id: I28bef8649783ae57ee94fe8b4d8590432e25c2fb