Improve handling of NEW_TASK deep links
When an external app starts a Navigation
deep link with NEW_TASK, we don't know
the state of the current task state, which
can lead to a confusing back button experience.

Ensure that we're in a predictably good state by
checking for Intents that have NEW_TASK but don't
have CLEAR_TASK and refire the Intent (via
TaskStackBuilder) with both flags.

NavDeepLinkBuilder was already using both flags, so
this only impacts web URI deep links. Updated the test
and testapp to take into account this behavior change.

BUG: 36817937
Test: NavControllerTest, manual testing with testapp
Change-Id: Ieea6a4a0f3c921c27fb82b670fb9514f35ff8b14