Rename tooltip related resources in support lib.
The layout file for the support library tooltip implementation
is named "tooltip.xml". Under some circumstances when an app using
the support library has a resource with the same name it gets loaded
instead and causes crashes.

Renaming it to abc_tooltip.xms, following the existing convention.
Doing the same for the two animation files (tooltip_enter/exit.xml).

Bug: 65147057
Test: Support7Demos shows tooltips on action bar buttons
Change-Id: I73f039718e48a6a5cdcea875fb12790f7cd5da4d