Don't process work multiple times.
This CL makes sure we don't try to process the same work item twice (this will
potentially happen if items A and B both have a constraint that is satisfied,
but B is enqueued after A... this will cause LiveData to fire with A two times).

Finally, this CL also makes sure ConstraintListeners are tracked in a Set and not
a List.  Otherwise you could add one each time the LiveData changed when a new
work item with that constraint was added.

Test: Update tests for all of these things.
Change-Id: I6a2e2c50cdd1e83e1eff9092e99069b892021585