Add Constraints to PeriodicWork; clean up logging.
Nothing in the current Constraints class actually conflicts with periodic work.
Also check to make sure we don't execute periodic work in the JobScheduler if
it was triggered due to an overrideDeadlineExpired.  We don't currently
support overrideDeadlines in PeriodicWork and it's not planned unless there is
a strong developer request for this.

Test: Existing tests.

Change-Id: I38f90a04747cdbfae5709f815ab6e3fa767c3eb6