Adds ProGuard processing into Jetifier.
We support all class spec and class filter rules. The most difficult part
of this was to poperly identify types referenced in the ProGuard files.
For this we are using a set of complex regular expression matchers. Once
the types are identified there are two ways how to rewrite them. If they
don't contain wildcards we take our TypesMap. If they have wildcards
inside we use ProGuardTypesMap that contains extended rules. Later we
can add capabaility to generate such map from existing rules.

Bug: b/64833662
Test: Check test/proguard package
Change-Id: I927b7243d641490bdd799210b86a47839044aaed