Mark many classes/methods as @RestrictTo(LIBRARY_SCOPE)/@hide.
Many classes unfortunately need to be public to accomodate our
class hierarchy and testing structure, but this will tell the IDE
to show a warning if the user tries to use them in their code.

This is not an exhaustive run through the files.  It covered the
main root directory, systemjob, utils, and firebase.  More to
come later.

Test: Ran existing tests.
Change-Id: Ifb0ae7f94d4b562c3286ebcb9485cccd424f2f4a