Cancel pending hover tooltip when updating tooltip text.
Currently if the tooltip text is updated just before a hover tooltip
is displayed, the update will be ignored and only become effective
after the tooltip is hidden and shown again.

Fixing this by tracking a pending tooltip and canceling it on an
update. This is safer than trying to update the tooltip in flight,
because it is known to lead to problems when view recycling is involved
(see http://b/38491655)

Bug: 68759508
Test: manual
Change-Id: I485543146ffa0cbcbfbd7222ba3543a0526780d5