Remove restriction for ItemCap extension.
Adapters that are set on the PagedListView no longer need to extend
ItemCap. Instead, if the adapters do not extend ItemCap, then setting
max number of pages will not do anything.

Also, update the tests to use Espresso calls instead of IdlingRegistry
because the latter does not work as reliably.

Test: Test on sample app and espresso tests
Bug: 69174396
Change-Id: If5a94883828ba5359fdc073a6f1c1be311792a32