Return a result from Worker#doWork().
Instead on relying on Exceptions to indicate failures, we now explicitly
return a result code from doWork.  This can be success, failure, or an
explicit retry.

Also, fixed the edge case where a processor is already executing some
work that another processor tries to do.  In this case, the second
processor should back off and reschedule that work for future execution.
Note that this may deviate the expected delay between the two processors;
this is expected.

Updated all the current Worker#doWork() methods.

Test: updated tests.

Change-Id: If851baf35543d00f475cbcfba049cee3926d70e5