No functional changes, just an effort to cleanup some code to decrease complexity.
Changes include:
1. LinearLayout.AnchorInfo is now static to make its relationship to
LinearLayout more clear.
2. RecyclerView defines a package private DEFAULT_ORIENTATION to more
clearly bind the shared default between the getProperties method and
the defaults expressed in LinearLayoutManager and
StaggeredLayoutManager.
3. LinearLayoutManager no longer lazily creates OrientationHelper such
that is it garunteed to be valid and nonnull after creation and inside
of LinearLayouts AnchorInfo (previously, AnchorInfo assumed it was non
null without much of a garuntee.
4. Now HORIZONTAL and VERTICAL are defined by LinearLayout -> RecyclerView
-> (everything else) instead of LinearLayout -> RecyclerView ->
-> OrientationHelper -> (everything else).

Bug: None.
Test: None.

Change-Id: I6edcc96383fc81f4c6450f58c6b83f543edc1956