Support for chaining arguments.
This CL introduces Worker.getArgumentsForChainedWork, which allows you to
pass an Arguments object downstream to Workers that depend on this one.
Arguments are merged into the Arguments specified when the Work object is
built, as well as any other Arguments provided by other parent Workers.
In case of a conflict, new values and types clobber old values and types.

Test: added tests.
Change-Id: I00c4031c3d4f6b559cac6ab997237d19f7c6d204