Make default emoji presentation style configurable
Certain emojis (such as U+2639 or U+26F9) are by default rendered with
the text presentation mode, unless followed by the variation selector
character U+FE0F. Unicode TR51 recommends that, if no variation selector is
present, the presentation style should be chosen according to the
nature of the application. For example, the emoji presentation style
would be desired in a messaging app, while a text editor would use the
text style.
This CL adds a configuration flag to EmojiCompat, which can be used to
make the emoji presentation style the default. If the flag is not set,
the emojis will use the text presentation style as before.

Bug: 67920701
Test: ./gradlew support-emoji:connectedAndroidTest
Change-Id: I10200e5ce5c56826a4710d49effad5e46481a9e0
(cherry picked from commit ca429f4bd64320e424b2e984c2e9d615679fa9cd)