Create WorkInput table.
WorkInput is a mapping of work id->Arguments, which will allow us to
potentially store multiple inputs for each Worker (in the case where
there are multiple prerequisites for a Work unit).

Right now, we are naively merging all of the input Arguments before
starting a job, but in the next CL, it will be done more intelligently.

One consequence of this CL is that we can no longer insert things in
workSpecDao only; we must also update workInputDao.  The changes in
tests reflect this.

Test: Updated tests.
Change-Id: Ifedb5c7264d4732e4b7cfed5713a9610b5d31221