Fix z-order of a tooltip on a view inside a Dialog.
Use the correct app window token for the tooltip popup (the way it is
already done in the platform implementation). This alone fixes z-order,
but changes the root view to which the tooltip offsets are applied,
and with it the observed position of the tooltip.

To fix that, change the root view finding logic. Always use the anchor's
root view directly for TYPE_APPLICATION windows. Only look through the
chain of contexts in case of non-app windows (such as popup windows or
sub panels).

Bug: 67623818
Test: manual
Change-Id: I3e212c3a6546739df4b9a1824556f07b3a36eab4