Start removing non-public stuff from base package.
Our base package exposes too much of our internals.  This is the
first pass at removing some of this stuff and moving it to an
impl subpackage.  Changes include:

- Move WorkManagerImpl to impl/
- Move configuration and initializer to impl/
- Create WorkContinuationImpl and make WorkContinuation abstract
- Move WorkDatabase to impl/
- Move BaseWork to impl/

Future cleanup: Work and PeriodicWork should probably get impls
(and revert the scope changes I made to move BaseWork).
Scheduler and Processor and a few other classes should move too
(saving that for another CL to make these manageable to review).

Test: Updated some tests
Change-Id: I2188e5ab60fb91891a80a8652bc83991c72f74be