Fix divider positioning logic.
When the divider was to be drawn relative to a view, it utilized methods
such getLeft() and getRight(). These methods returned a position
relative to that View's parent. This parent is not necessarily the
container and thus, the divider would be drawn at an incorrect position.

Instead, utilize getGlobalVisibleRect(), which returns a bounding box of
the view in global coordinates. Then, compare these coordinates with
that of the container to retrieve the relative positioning.

Test: Test on sample app.
Bug: 70354251
Change-Id: I335bbfe49d4d3761b990ac8092863951a2603c9e