[webkit-codegen] Add code generation for boundary interfaces.
Add some code for generating boundary interfaces (the interfaces used
for communication between the support library and the WebView APK)
given a representation of a class and the methods we should generate
interfaces for.

Currently, the WebView, WebViewProvider, WebViewFactoryProvider, and
inner classes are not supported. All other classes should be fine.

We will add copyright generation and use javac to compile the generated
test output files (to ensure they're valid) in follow-up CLs.

Bug: 70459808
Test: run added tests.
Change-Id: I14245c23896b21003e88c1bf5b121cbfc1b66abd