Fixed the failing test and reorganized the test a bit to make the code paths a bit clearer.
The primary issue with the test was that on API 17 (and probably 19), the FastScroller hadn't
yet retrieved the dimensions of the RecyclerView because it does so within onDraw.  I've
updated the test to not proceed until at least the first onDraw event has occured.

Further cleanup of the test would be nice, but that is for future work.

Bug: 69873706
Test: RecyclerViewFastScrollerTest#sanityScrollingInstrumentation

Change-Id: I2f394530299d811621650b4eb738e83a6cdbcf12