Get work statuses for an enqueued chain.
There are instances where identifying work by just id
and tag is not sufficient.  In those situations, it's
more fluent to be able to observe the status of all
the work you just enqueued in a chain.  This CL adds
WorkContinuation#getStatuses, which returns a LiveData
with the payload as a Map of WorkSpec ids to their
WorkStatuses.

Test: Added and ran tests.
Change-Id: Iad9bef7dc11391ed738fc0fb32a98f53dc7f6ef5