Allow collection type converters in query parameters
If a query parameter is a collection AND we cannot find a converter for the type parameter of it,
we will look for a converter that converts the whole thing.

Collection Type converters were broken in kotlin data classes because
kotlin creates constructor args with variance which cannot be assigned
to the fields. This CL flexes the constructor and setter check to allow
variances.

Fixes: 69164099
Test: SimpleEntityReadWriteTest, CustomTypeConverterResolutionTest, BookDaoTest,
TypeAssignmentTest

Change-Id: I48d44bb90c69ad0d9ca4aeb19f22821d89264405