Allow single column results in relations
This CL fixes a bug where we were enforcing @Relation's return type
to adhere to the Pojo requirements which breaks if you want to use
a boxed primitive or a type with type converter.

This CL changes relationship processor to not enforce a pojo as long as
we can infer the projection from available information (e.g. user spcified).

If we need to infer projection list, we now first check if there is
a column adapter for the given type and if so, assume it is a 1 column
result and generate query accordingly.

Fixes: 68077506
Test: BooksDaoTest, PojoProcessorTest, PojoWithRelationTest
Change-Id: I34484c11248ba9be0341f498b55ecbf848f8a35b