Fix schema identity has to be stable
Room v1 has a bug where it generates the schema identity from the
create table query, which was OK when it was written (because Room
considered column order as part of schema) but not anymore (since
we don't enforce column order).

The problem is fixed by creating both legacy and new identity hash
so that RoomOpenHelper can validate old database.

Unfortunately, this is a large CL because we don't want to update
schema json files so all of the Schema data classes implements a
new schema equality api which checks if the schema description of
two entities are the same, even though their SQL might be different.
(e.g. the column order in an entity or the auto generated index name)

Since we are not overriding json files, the fix will only take effect
after a migration.

Bug: 64290754
Test: android.arch.persistence.room.migration.bundle.*
Change-Id: I44959a353ac919850e2606ca704008ea57da7313